IRISHUB-359, IRISHUB-360: move input.go and keys.go to client/keys
