+++ client/keys/keys.go
@@ -1,4 +1,4 @@
-package client
+package keys
 
 import (
 	"github.com/cosmos/cosmos-sdk/crypto/keys"
@@ -7,7 +7,7 @@ import (
 
 // GetKeyBase initializes a keybase based on the given db.
 // The KeyBase manages all activity requiring access to a key.
-func GetKeyBase(db dbm.DB) keys.Keybase {
+func GetKeyBaseFromDB(db dbm.DB) keys.Keybase {
 	keybase := keys.New(
 		db,
 	)
@@ -18,5 +18,5 @@ func GetKeyBase(db dbm.DB) keys.Keybase {
 // useful for --dry-run to generate a seed phrase without
 // storing the key
 func MockKeyBase() keys.Keybase {
-	return GetKeyBase(dbm.NewMemDB())
+	return GetKeyBaseFromDB(dbm.NewMemDB())
 }
